LastChanges

+++ API/src/controllers/todo.js
@@ -3,16 +3,21 @@ import Todo from "../models/todo.js";
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
-import fs from 'fs/promises';
+//import fs from 'fs/promises';
+import fs from 'fs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-export const getAllByEmail = async (req, res, next) => {
-    const userEmail = req.params.email;
+export const getAllByUser = async (req, res, next) => {
+    const userId = req.params.userId;
 
     try {
-      const todos = await Todo.find({ user: userEmail });
+      const todos = await Todo.find({ userId: userId });
+
+
+      if (todos.length > 0  && todos[0].userId !== req.userId)  return res.status(403).send("You are not allowed!");
+      
       res.status(200).json(todos);
     } catch (error) {
       res.status(500).json({ message: error.message });
@@ -20,13 +25,18 @@ export const getAllByEmail = async (req, res, next) => {
   };
 
 export const add = async (req, res, next) => {
-    const { user, todo, tag } = req.body; 
-    const img = req.files['thumbnail'] ? req.files['thumbnail'][0].filename : '';
-    const file = req.files['file'] ? req.files['file'][0].filename : '';
+  console.log(req.body);
+    const { userId, todo, tag } = req.body; 
+    let img, file;
+    if (req.files != undefined ) {
+      img = req.files['thumbnail'] ? req.files['thumbnail'][0].filename : '';
+      file = req.files['file'] ? req.files['file'][0].filename : '';
+    }
     
-    try {
+    
+    //try {
         const newTodo = new Todo({
-        user,
+        userId,
         img,
         file,
         todo,
@@ -36,9 +46,9 @@ export const add = async (req, res, next) => {
         const savedTodo = await newTodo.save();
 
         res.status(201).json(savedTodo);
-    } catch (error) {
-        res.status(500).json({ message: error.message });
-    }
+    //} catch (error) {
+      //  res.status(500).json({ message: error.message });
+    //}
 };
 
 export const deleteById = async (req, res, next) => {
@@ -51,15 +61,21 @@ export const deleteById = async (req, res, next) => {
       return res.status(404).json({ message: 'Todo not found' });
     }
 
-    if (todoToDelete.img !== '') {
-      const imgPath = path.join(__dirname, '../../uploads', todoToDelete.img);
-      await fs.unlink(imgPath); 
-    }
+    if (todoToDelete.userId !== req.userId)  return res.status(403).send("You are not allowed!");
 
-    if (todoToDelete.file !== '') {
-      const filePath = path.join(__dirname, '../../uploads', todoToDelete.file);
-      await fs.unlink(filePath);
+    try {
+      if (todoToDelete.img !== '') {
+        const imgPath = path.join(__dirname, '../../uploads', todoToDelete.img);
+        await fs.unlink(imgPath); 
+      }
+      if (todoToDelete.file !== '') {
+        const filePath = path.join(__dirname, '../../uploads', todoToDelete.file);
+        await fs.unlink(filePath);
+      }
+    } catch {
+      console.log('files do no exits');
     }
+    
     await Todo.findByIdAndDelete(_id);
 
     res.status(200).json({ message: 'Todo and associated files deleted successfully' });
@@ -97,10 +113,9 @@ export const downloadFile = async (req, res, next) => {
   try {
     
     const fileName = req.params.fileName;
-    console.log("downloaddding:", fileName);
     
     const filePath = path.join(__dirname, '../../uploads/', fileName);
-    console.log(filePath);
+
     res.download(filePath, (err) => {
       if (err) {
         res.status(404).json({ message: 'File not found or error downloading' });
@@ -112,10 +127,10 @@ export const downloadFile = async (req, res, next) => {
 };
 
 export const search = async (req, res, next) => {
-  const { key } = req.params;
+  const { key, userId } = req.params;
   
   try {
-    const todos = await Todo.find({ todo: { $regex: key, $options: 'i' } });
+    const todos = await Todo.find({ userId: userId, todo: { $regex: key, $options: 'i' } });
     res.json(todos);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -132,11 +147,41 @@ export const getAllTags = async (req, res, next) => {
 };
 
 export const getByTag = async (req, res, next) => {
-  const { tag } = req.params;
+  const { tag, userId } = req.params;
   try {
-    const todos = await Todo.find({ tag: tag });
+    const todos = await Todo.find({ userId: userId, tag: tag });
     res.status(200).json(todos);
   } catch (error) {
     res.status(500).json({ message: error.message });
+  }
+};
+
+export const uploads = async (req, res, next) => {
+  try {
+    console.log("asdadada");
+    const imgName = req.params.img;
+    const imgPath = path.join(__dirname, '..', '..', 'uploads', imgName);
+
+    // Check if the file exists
+    if (fs.existsSync(imgPath)) {
+      // Set the appropriate content type
+      const ext = path.extname(imgName).toLowerCase();
+      const contentType = {
+        '.png': 'image/png',
+        '.jpg': 'image/jpeg',
+        '.jpeg': 'image/jpeg',
+      }[ext] || 'application/octet-stream';
+
+      res.setHeader('Content-Type', contentType);
+
+      // Stream the file
+      const stream = fs.createReadStream(imgPath);
+      stream.pipe(res);
+    } else {
+      res.status(404).send('Image not found');
+    }
+  } catch (error) {
+    console.error('Error serving image:', error);
+    res.status(500).send('Internal Server Error');
   }
 };
